Backend-main/massuh-datalake-main/api-gateway/requirements.txt
functions-framework==3.*
pandas
google-cloud-bigquery
google-auth
pandas-gbq
